Implement getAlises in core

+++ src/rebostCore.py
@@ -355,7 +355,8 @@ class _RebostCore():
 				for f in os.scandir(raw):
 					os.unlink(f.path)
 		self.stores["main"].remove_all()
-		self.stores["mainB"].remove_all()
+		if "mainB" in self.stores:
+			self.stores["mainB"].remove_all()
 		for storeId in self.stores.keys():
 			self._debug("Process {}".format(storeId))
 			if isinstance(storeId,int):
@@ -450,7 +451,31 @@ class _RebostCore():
 		return(self.config.get("externalInstaller",""))
 	#def getExternalInstaller
 
-	def _getRemoteAppFixes(self,url):
+	def _chkRemoteFileChanged(self,url):
+		changed=False
+		fname=url.split("://")[-1]
+		fname=fname.replace("/","_")
+		fTag=os.path.join(CACHE,"{}.etag".format(fname))
+		eTag=""
+		oldETag=""
+		try:
+			headers=Request(url,method="HEAD",headers={'User-Agent':'Mozilla/5.0'})
+			with urlopen(headers,timeout=2) as f:
+				eTag=f.headers.get("ETag")
+		except:
+			eTag=""
+		if os.path.exists(fTag):
+			with open(fTag,"r") as f:
+				oldETag=f.read()
+			self._debug("Check ETag: {} {}".format(eTag,oldETag))
+		if eTag!=oldETag:
+			changed=True
+			with open(fTag,"w") as f:
+				f.write(eTag)
+		return(changed)
+	#def _chkRemoteFileChanged
+
+	def _getRemoteMapFixes(self,url):
 		content={}
 		urlContent=""
 		if url!="" and isinstance(url,str):
@@ -468,7 +493,19 @@ class _RebostCore():
 				print(e)
 				content={}
 		return(content)
-	#def _getRemoteAppFixes
+	#def _getRemoteMapFixes
+
+	def _getLocalMapFixes(self,fpath):
+		content={}
+		if os.path.exists(fpath):
+			with open(fpath,"r") as f:
+				try:
+					content=json.loads(f.read())
+				except Exception as e:
+					print("ERROR PROCESSING {}",format(fpath))
+					print(e)
+		return(content)
+	#def _getLocalMapFixes
 
 	def _getLocalMapFiles(self):
 		#Load map files from DATA
@@ -494,25 +531,36 @@ class _RebostCore():
 		return(mapFiles)
 	#def _getLocalMapFiles
 
+	def _getCacheMapFiles(self):
+		mapFiles=[]
+		for f in os.scandir(CACHE):
+			if f.name.endswith(".map"):
+				mapFiles.append(f.path)
+		return(mapFiles)
+	#def _getCacheMapFiles
+
 	def getMapFixes(self):
 		mapFixes={"nodisplay":[],"aliases":{}}
 		mapFiles=self._getLocalMapFiles()
 		for mapFile in mapFiles:
 			self._debug("Reading mapF {}".format(mapFile))
 			if os.path.exists(mapFile) and mapFile.endswith(".map"):
-				mapFixesF={}
-				with open(mapFile,"r") as f:
-					try:
-						mapFixesF=json.loads(f.read())
-					except Exception as e:
-						print("ERROR PROCESSING {}",format(mapFile))
-						print(e)
+				mapFixesF=self._getLocalMapFixes(mapFile)
 				rContent={}
 				if "upstream" in mapFixesF:
 					self._debug("Checking remote URL {}".format(mapFixesF["upstream"]))
-					rContent=self._getRemoteAppFixes(mapFixesF["upstream"])
-					rContent["upstream"]=mapFixesF["upstream"]
-				if len(rContent["nodisplay"])>0:
+					fname=mapFixesF["upstream"].split("://")[-1]
+					fname=fname.replace("/","_")
+					fcache=os.path.join(CACHE,"{}".format(fname))
+					if self._chkRemoteFileChanged(mapFixesF["upstream"])==True or os.path.exists(fcache)==False:
+						rContent=self._getRemoteMapFixes(mapFixesF["upstream"])
+						rContent["upstream"]=mapFixesF["upstream"]
+						with open(fcache,"w") as f:
+							f.write(json.dumps(rContent,indent=4))
+					else:
+						self._debug("Getting mapping from cache")
+						rContent=self._getLocalMapFixes(fcache)
+				if len(rContent.get("nodisplay",[]))>0:
 					mapFixesF=rContent.copy()
 					with open(mapFile,"w") as f:
 						f.write(json.dumps(mapFixesF,indent=4))
@@ -520,6 +568,19 @@ class _RebostCore():
 				mapFixes["aliases"].update(mapFixesF.get("aliases",{}))
 		return mapFixes
 	#def getMapFixes
+
+	def chkCacheNeedsUpdate(self):
+		update=False
+		cFiles=self._getCacheMapFiles()
+		for cFile in cFiles:
+			with open(cFile,"r") as f:
+				content=json.loads(f.read())
+			if "://" in content.get("upstream",""):
+				if self._chkRemoteFileChanged(content["upstream"])==True:
+					update=True
+					break
+		return update
+	#def chkCacheNeedsUpdate
 
 	def initCore(self):
 		#self.thExecutor.submit(self._loadFromCache)
